Add files via upload
If running a 'Receiver' test, load the terminal at that folder. The current OFFSPRING PERSISTENCE FILES are from a MOTHER initial run which was done on the author's machine as a replicable start up condition.

If running a 'Master' test, note that the audio output is set for the 'Audient' DAC, a 'prosumer' audio interface for audio recording from the company Audient. Another USB DAC can be substituted for 'OUTPUT' but it is recommended to keep the 'INPUT' the same, so, on the local machine's default.